added ability to handle multiple versions of a survey, and to update surveys even when their version numbers don't change. past responses and reminders are tied to a survey by schema namespace and schema name so even if the survey content or version number changes, response history and reminders are preserved.
a survey now stores an SHA1 hash of the JSON dictionary definition that created it, and a new version of the survey is created whenever the hashes don't match.

added survey version text to survey detail and survey response screens so user can see if they're looking at a response to an old version (if the version number was updated).

also made it so user is logged out and sign-in screen is presented when database is reset since db was modified to add SHA1 hash to surveys.